Fixes #440 (Unnecessary nonlinear equation systems with Fluid.Sources):
Modelica.Fluid.Sources.FixedBoundary/.Boundary_pT/.Boundary_ph:

Changed the implementation so that no non-linear algebraic equation system occurs, if the given variables (e.g. p,T,X) do not correspond to the medium states (e.g. p,h,X). This is achieved by using appropriate "setState_xxx" calls to compute the medium state from the given variables. If a nonlinear equation system occurs, it is solved by a specialized handler inside the setState_xxx(..) function, but in the model this equation system is not visible.

Added this improvement to the release notes.