#323 fixed (Check usage of regRoot2 throughout the library):
The final issue was that Dymola complains with the pedantic flag:
  "Differentiating xxx under the assumption that it is continuous at switching."

This seems to be too pedantic, since the function in which this expression is used
has an annotation "smoothOrder", and the tools should just accept this definition 
from the user.

In order to avoid this error message, a "smooth(2, ...)" operator has been 
introduced around the corresponding expression.