This commit fixes a number of tickets at once  - #924, which was not a single mistake but the same problem existed all over the place. Fixing #924 required moving a number of types into separate, non-partial packages, mostly to Media.Interfaces.Types.  - #957 was fixed by giving defaults (false in both cases) to the package constants onePhase and smoothModel  - #649 was fixed by re-declaring sizes in a number of functions in PartialTwoPhaseMedium from [nX], which is wrong for some settings, to [:].
In addition, #799 was fixed for all problems in Modelica.Media. Also, the test models used inside Media were declared deprecated (they were still using semiLinear), and replaced by references to Modelica.Fluid.

The last move introduces a reference to Modelica.Fluid in Modelica.Media, which is a sort of circular reference. This will be flagged in a seprate ticket, and the models should be moved out out of Modelica.Media to MedelicaTest. The problem is that this _may_ be considered a non-backwards compatible change. I don't think so for models which only had the purpose to be test models.